Update feed, minor restructure
Users can now update their feed posts which completes CRUD for this "product." I'm not sure if this is required because I did have all CRUD elements before, but not all for the product. Youtube links still need to be verified and I need more comments, but other than that this milestone is about complete.