Milestone 6 Complete
This will most likely be the final push of milestone 6, but there is a bit of bug testing that still needs to be done, so it could change.

Changes with this update include: ajax form posting for minesweeper game, account editor, and feed post update/delete. No more page refreshes for any of those. 500 million javadoc style comments replaced normal people comments throughout code. Sorry might be a bit salty about the amount of time that took <.<