Like dislike Feed system, Dating removed from home
This is a small update that adds likes and dislikes to the feed post system and removes the dating mini-dashboard from the user home page (because I forgot). The voting system is being done in lieu of comments with this project due to time constraints and more important functionality to be met before presentation. Comments may still be implemented if time permits, but I doubt it because i really want to do sodoku and that's at the end of the list.