User profiles, more detailed search results
Reaching the final wrap up of this project. Today, I finished up the search list and started on the profile viewer. Currently, the profile viewer is bugged and will take you to the feed list page if you like, dislike, remove, or update a feed post. That's because this system was copy-pasted from the feed list. This will be fixed tomorrow and is just being used as a placeholder for the time being. I hope to finish the friend request system and message system tomorrow to wrap up the functionality; however, there is a small possibility of this not happening. Due to the complexity of the friend request system and message system, these may be delayed to the senior project. Worst case scenario, view profile will function as expected after tomorrow.